fix: Changing phone_number type from integer to bigint

+++ db/migrate/20240920115823_add_phone_number_to_users.rb
@@ -1,5 +1,5 @@
 class AddPhoneNumberToUsers < ActiveRecord::Migration[7.1]
   def change
-    add_column :users, :phone_number, :integer
+    add_column :users, :phone_number, :bigint
   end
 end
